Make Data and SerializableData interfaces
